server/requirements.txt
flask
flask-cors
flask-pymongo
python-dotenv
werkzeug
bcrypt
PyJWT
# Video and Audio Processing
opencv-python==4.8.1.78
mediapipe==0.10.8
protobuf==3.20.3
ffmpeg-python==0.2.0
# Audio Analysis
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
# Speech Recognition
openai-whisper==20231117
# Text Analysis
nltk==3.8.1
textstat==0.7.3
# LLM for Feedback
openai==1.3.0
# Utilities
pydub==0.25.1
pillow==10.1.0